Decomposition of project
