+++ lab1/Test/Test.h
@@ -1,8 +1,8 @@
 #pragma once
 #include <vector>
 #include "../Test/unit_test_framework.h"
-#include "../lab1/HeaderFiles/Containers/Containers.h"
-#include "../lab1/HeaderFiles/Date/Date.h"
+#include "../lab1/Containers/Containers.h"
+#include "../lab1/Work with date/WorkWithDate.h"
 
 void TestStack();
 void TestQueue();
