+++ lab1/lab1/Work with date/WorkWithDate.h
@@ -1,0 +1,12 @@
+/**
+*	File which linked "Date", "DifferenceBetweenDate", "TimeIsNow" and "TimeManagment" libraries.
+*   @file WorkWithDate.h
+*	@author Kirill Kotsko
+*	@version 1.2
+*/
+
+#pragma once
+#include "Date.h"
+#include "DifferenceBetweenDate.h"
+#include "TimeIsNow.h"
+#include "TimeManagment.h"
