+++ lab1/lab1/lab1.vcxproj.filters
@@ -13,69 +13,66 @@
       <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
       <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
     </Filter>
-    <Filter Include="CppFiles">
-      <UniqueIdentifier>{720ba3fb-6950-4cc7-84cf-13130d55d82e}</UniqueIdentifier>
+    <Filter Include="UserInterface">
+      <UniqueIdentifier>{193deb28-0e32-428b-80fd-ab5d62b1cf00}</UniqueIdentifier>
     </Filter>
-    <Filter Include="HeaderFiles">
-      <UniqueIdentifier>{613d9a78-7205-48a0-995e-351657eb2f6a}</UniqueIdentifier>
+    <Filter Include="Work with date">
+      <UniqueIdentifier>{89d94dc6-1bfe-4224-8fc5-74c0b7ac7f8c}</UniqueIdentifier>
     </Filter>
-    <Filter Include="HeaderFiles\Containers">
-      <UniqueIdentifier>{e078647e-bdbc-45ed-91f3-1507a907e592}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="HeaderFiles\Date">
-      <UniqueIdentifier>{0014739c-9717-4958-9823-b4fffbff293f}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="HeaderFiles\Demonstration">
-      <UniqueIdentifier>{6ebe5b50-c24a-4c4e-b809-41662844b158}</UniqueIdentifier>
+    <Filter Include="Containers">
+      <UniqueIdentifier>{6cff81a5-3bd7-48b0-9502-45a92c09c1a4}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="main.cpp">
       <Filter>Исходные файлы</Filter>
     </ClCompile>
-    <ClCompile Include="CppFiles\Overload.cpp">
-      <Filter>CppFiles</Filter>
+    <ClCompile Include="UserInterface\UI.cpp">
+      <Filter>UserInterface</Filter>
     </ClCompile>
-    <ClCompile Include="CppFiles\Time_is_now.cpp">
-      <Filter>CppFiles</Filter>
+    <ClCompile Include="Work with date\Date.cpp">
+      <Filter>Work with date</Filter>
     </ClCompile>
-    <ClCompile Include="CppFiles\Time_managment.cpp">
-      <Filter>CppFiles</Filter>
+    <ClCompile Include="Work with date\DifferenceBetweenDate.cpp">
+      <Filter>Work with date</Filter>
     </ClCompile>
-    <ClCompile Include="CppFiles\UI_menu.cpp">
-      <Filter>CppFiles</Filter>
+    <ClCompile Include="Work with date\TimeIsNow.cpp">
+      <Filter>Work with date</Filter>
     </ClCompile>
-    <ClCompile Include="CppFiles\Demo_stack.cpp">
-      <Filter>CppFiles</Filter>
-    </ClCompile>
-    <ClCompile Include="CppFiles\Demo_queue.cpp">
-      <Filter>CppFiles</Filter>
-    </ClCompile>
-    <ClCompile Include="CppFiles\Demo_deque.cpp">
-      <Filter>CppFiles</Filter>
-    </ClCompile>
-    <ClCompile Include="CppFiles\Time_managment_operation.cpp">
-      <Filter>CppFiles</Filter>
+    <ClCompile Include="Work with date\TimeManagment.cpp">
+      <Filter>Work with date</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="HeaderFiles\Containers\Containers.h">
-      <Filter>HeaderFiles\Containers</Filter>
+    <ClInclude Include="Work with date\Date.h">
+      <Filter>Work with date</Filter>
     </ClInclude>
-    <ClInclude Include="HeaderFiles\Containers\Stack.h">
-      <Filter>HeaderFiles\Containers</Filter>
+    <ClInclude Include="Work with date\DifferenceBetweenDate.h">
+      <Filter>Work with date</Filter>
     </ClInclude>
-    <ClInclude Include="HeaderFiles\Containers\Queue.h">
-      <Filter>HeaderFiles\Containers</Filter>
+    <ClInclude Include="Work with date\TimeIsNow.h">
+      <Filter>Work with date</Filter>
     </ClInclude>
-    <ClInclude Include="HeaderFiles\Containers\Deque.h">
-      <Filter>HeaderFiles\Containers</Filter>
+    <ClInclude Include="Work with date\TimeManagment.h">
+      <Filter>Work with date</Filter>
     </ClInclude>
-    <ClInclude Include="HeaderFiles\Date\Date.h">
-      <Filter>HeaderFiles\Date</Filter>
+    <ClInclude Include="UserInterface\UI.h">
+      <Filter>UserInterface</Filter>
     </ClInclude>
-    <ClInclude Include="HeaderFiles\Demonstration\Demonstration.h">
-      <Filter>HeaderFiles\Demonstration</Filter>
+    <ClInclude Include="Containers\Containers.h">
+      <Filter>Containers</Filter>
+    </ClInclude>
+    <ClInclude Include="Containers\Deque.h">
+      <Filter>Containers</Filter>
+    </ClInclude>
+    <ClInclude Include="Containers\Queue.h">
+      <Filter>Containers</Filter>
+    </ClInclude>
+    <ClInclude Include="Containers\Stack.h">
+      <Filter>Containers</Filter>
+    </ClInclude>
+    <ClInclude Include="Work with date\WorkWithDate.h">
+      <Filter>Work with date</Filter>
     </ClInclude>
   </ItemGroup>
 </Project>
