+++ lab1/lab1/main.cpp
@@ -4,9 +4,9 @@
 *	@version 1.1
 */
 
-#include "../lab1/HeaderFiles/Containers/Containers.h"
-#include "../lab1/HeaderFiles/Date/Date.h"
-#include "../lab1/HeaderFiles/Demonstration/Demonstration.h"
+#include "../lab1/Containers/Containers.h"
+#include "../lab1/Work with date/Date.h"
+#include "../lab1/UserInterface/UI.h"
 
 using namespace std;
